+++ app/Models/DailyRawMetricsModel.php
@@ -30,9 +30,9 @@ class DailyRawMetricsModel extends AbstractModel
         $this->shema()->create($name, function ($table) {
             /** @var \Illuminate\Database\Schema\Blueprint $table */
             $table->increments('id');
-            $table->unsignedSmallInteger('metric_id');
-            $table->integer('value');
-            $table->unsignedSmallInteger('minute');
+            $table->unsignedInteger('metric_id');
+            $table->unsignedBigInteger('value');
+            $table->unsignedInteger('minute');
         });
     }
