+++ app/Models/DailyRawSlicesModel.php
@@ -32,10 +32,10 @@ class DailyRawSlicesModel extends AbstractModel
         $this->shema()->create($name, function ($table) {
             /** @var \Illuminate\Database\Schema\Blueprint $table */
             $table->increments('id');
-            $table->unsignedSmallInteger('metric_id');
-            $table->unsignedSmallInteger('slice_id');
-            $table->integer('value');
-            $table->unsignedSmallInteger('minute');
+            $table->unsignedInteger('metric_id');
+            $table->unsignedInteger('slice_id');
+            $table->unsignedBigInteger('value');
+            $table->unsignedInteger('minute');
         });
     }
